+++ mail/mail-uniffi/CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - [ET-1987] Fix email address validation for `RecipientList`.
   - [ET-1923] Remove mail settings signature.
   - [ET-1913] Support retrying of failed login flow stages
+  - Fix sender address repeated in To and CC on draft reply.
 
 ### Changed